+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.7.42
 
+* fix: fix missing source after cleaning layout elements
 * Remove chipper model
 
 ## 0.7.41